requirements.txt
adlfs==2025.8.0
aiobotocore==2.24.2
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiohttp-retry==2.9.1
aioitertools==0.12.0
aiooss2==0.2.11
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
amqp==5.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.10.0
appdirs==1.4.4
argcomplete==3.6.2
asyncssh==2.21.0
atpublic==6.0.1
attrs==25.3.0
azure-core==1.35.1
azure-datalake-store==0.0.53
azure-identity==1.25.0
azure-storage-blob==12.26.0
bcrypt==4.3.0
billiard==4.2.1
boto3==1.40.18
botocore==1.40.18
cachetools==5.5.2
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
configobj==5.0.9
contourpy==1.3.3
crcmod==1.7
cryptography==43.0.3
cycler==0.12.1
decorator==5.2.1
dictdiffer==0.9.0
diskcache==5.6.3
distro==1.9.0
dpath==2.2.0
dulwich==0.24.1
dvc==3.63.0
dvc-azure==3.1.0
dvc-data==3.16.12
dvc-gdrive==3.0.1
dvc-gs==3.0.2
dvc-hdfs==3.0.0
dvc-http==2.32.0
dvc-objects==5.1.1
dvc-oss==3.0.0
dvc-render==1.0.2
dvc-s3==3.2.2
dvc-ssh==4.2.1
dvc-studio-client==0.22.0
dvc-task==0.40.2
dvc-webdav==3.0.0
dvc-webhdfs==3.1.0
entrypoints==0.4
filelock==3.19.1
flatten-dict==0.4.2
flufl.lock==8.2.0
fonttools==4.60.0
frozenlist==1.7.0
fsspec==2025.9.0
funcy==2.0
gcsfs==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-core==2.4.3
google-cloud-storage==3.4.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grandalf==0.8
gto==1.8.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hydra-core==1.3.2
idna==3.10
isodate==0.7.2
iterative-telemetry==0.0.10
jmespath==0.10.0
joblib==1.5.2
kiwisolver==1.4.9
knack==0.12.0
kombu==5.5.4
markdown-it-py==4.0.0
matplotlib==3.10.6
mdurl==0.1.2
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
networkx==3.5
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
omegaconf==2.3.0
orjson==3.11.3
oss2==2.18.4
ossfs==2025.5.0
packaging==25.0
pandas==2.3.2
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.0.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.11.9
pydantic_core==2.33.2
pydot==4.0.1
PyDrive2==1.21.3
pygit2==1.18.2
Pygments==2.19.2
pygtrie==2.5.0
PyJWT==2.10.1
pyOpenSSL==24.2.1
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.1.0
rsa==4.9.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
s3fs==2025.9.0
s3transfer==0.13.1
scikit-learn==1.7.2
scipy==1.16.2
scmrepo==3.5.2
seaborn==0.13.2
semver==3.0.4
setuptools==80.9.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqltrie==0.11.2
sshfs==2025.2.0
tabulate==0.9.0
threadpoolctl==3.6.0
tomlkit==0.13.3
tqdm==4.67.1
typer==0.17.4
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
voluptuous==0.15.2
wcwidth==0.2.13
webdav4==0.10.0
wrapt==1.17.3
yarl==1.20.1
zc.lockfile==3.0.post1
